fix: Change ratePerMile type from int to double in DeliveryFeeDeliverySettings

+++ lib/src/domain/checkout/models/calculated_delivery_charge_details_model.dart
@@ -421,7 +421,7 @@ class DeliveryFeeDeliverySettings {
   final int? calibratedDistance;
   final double? freeDeliveryRadius;
   final double? maxDeliveryRadius;
-  final int? ratePerMile;
+  final double? ratePerMile;
   final String? deliveryChargeType;
   final int? freeDeliveryMinOrder;
   final String? calculateNormalDeliveryFee;
@@ -444,7 +444,7 @@ class DeliveryFeeDeliverySettings {
     int? calibratedDistance,
     double? freeDeliveryRadius,
     double? maxDeliveryRadius,
-    int? ratePerMile,
+    double? ratePerMile,
     String? deliveryChargeType,
     int? freeDeliveryMinOrder,
     String? calculateNormalDeliveryFee,
@@ -499,7 +499,7 @@ class DeliveryFeeDeliverySettings {
               : map["maxDeliveryRadius"]
           : null,
       ratePerMile:
-          map['ratePerMile'] != null ? map['ratePerMile'] as int : null,
+          map['ratePerMile'] != null ? map['ratePerMile'] as double : null,
       deliveryChargeType: map['deliveryChargeType'] != null
           ? map['deliveryChargeType'] as String
           : null,
